refactor: generate qr code in web socket server side
