+++ app/[locale]/error.tsx
@@ -2,11 +2,11 @@
 
 import { useTranslations } from "next-intl";
 
-export default function Error() {
+export default function Error({ error }: { error: string }) {
   const t = useTranslations();
   return (
     <p style={{ color: "rgb(255, 85, 85)" }} className="font-bold text-xl">
-      {t("somethingWentWrong")}
+      {`${t("somethingWentWrong")}: ${error}`}
     </p>
   );
 }
