+++ app/hook/useSocket.ts
@@ -1,14 +1,18 @@
 import { useEffect, useState } from "react";
-import io, { Socket } from "socket.io-client";
+import io from "socket.io-client";
 import { transformClaimsData } from "../utils";
 import {
   OriginalPresentationEventMessage,
   PresentationEventMessage,
+  QRRequestState,
 } from "@/app/lib/definitions";
 
 export const useSocket = () => {
-  const [socket, setSocket] = useState<Socket | null>(null);
   const [isConnected, setIsConnected] = useState(false);
+  const [requestQRState, setRequestQRState] = useState<QRRequestState>({
+    loading: true,
+  });
+
   const [presentationEventMessage, setPresentationEventMessage] =
     useState<PresentationEventMessage>();
 
@@ -17,10 +21,26 @@ export const useSocket = () => {
 
     socketIo.on("connect", () => {
       setIsConnected(true);
+      socketIo.emit("generateQR", { socketConnectionId: socketIo.id });
     });
 
     socketIo.on("disconnect", () => {
       setIsConnected(false);
+    });
+
+    socketIo.on("generateQREventMessage", (msg) => {
+      console.log("generateQREventMessage", msg);
+      if (msg.ok && msg.shortUrl) {
+        setRequestQRState({
+          loading: false,
+          shortUrl: msg.shortUrl,
+        });
+      } else {
+        setRequestQRState({
+          loading: false,
+          error: msg.error ?? "An error occurred",
+        });
+      }
     });
 
     socketIo.on(
@@ -34,10 +54,8 @@ export const useSocket = () => {
           const { ref, status, proofExchangeId } = msg;
           setPresentationEventMessage({ ref, status, proofExchangeId });
         }
-      },
+      }
     );
-
-    setSocket(socketIo);
 
     return () => {
       socketIo.disconnect();
@@ -47,6 +65,6 @@ export const useSocket = () => {
   return {
     isConnected,
     presentationEventMessage,
-    socketConnectionId: socket?.id,
+    requestQRState,
   };
 };
