+++ server.js
@@ -4,11 +4,27 @@ const { parse } = require("url");
 const next = require("next");
 const { Server } = require("socket.io");
 
+const SERVICE_AGENT_ADMIN_BASE_URL =
+  process.env.SERVICE_AGENT_ADMIN_BASE_URL ||
+  "https://a.chatbot-demo.dev.2060.io";
+
+const CREDENTIAL_DEFINITION_ID =
+  process.env.CREDENTIAL_DEFINITION_ID ||
+  "did:web:chatbot-demo.dev.2060.io?service=anoncreds&relativeRef=/credDef/HngJhYMeTLTZNa5nJxDybmXDsV8J7G1fz2JFSs3jcouT";
+
 const dev = process.env.NODE_ENV !== "production";
 const app = next({ dev });
 const handle = app.getRequestHandler();
 
 app.prepare().then(() => {
+  const PORT = process.env.NEXT_PUBLIC_PORT
+    ? Number(process.env.NEXT_PUBLIC_PORT)
+    : 3000;
+
+  const PUBLIC_BASE_URL = process.env.NEXT_PUBLIC_BASE_URL
+    ? `${process.env.NEXT_PUBLIC_BASE_URL}:${PORT}`
+    : `http://localhost:${PORT}`;
+
   const server = createServer((req, res) => {
     const parsedUrl = parse(req.url, true);
     handle(req, res, parsedUrl);
@@ -18,7 +34,41 @@ app.prepare().then(() => {
 
   io.on("connection", (socket) => {
     console.log("A client connected info is:", socket.id);
-
+    let message = {};
+    socket.on("generateQR", async (data) => {
+      try {
+        const url = `${SERVICE_AGENT_ADMIN_BASE_URL}/v1/invitation/presentation-request`;
+        const requestBody = {
+          callbackUrl: `${PUBLIC_BASE_URL}/api/presentation`,
+          ref: data.socketConnectionId,
+          requestedCredentials: [
+            { credentialDefinitionId: CREDENTIAL_DEFINITION_ID },
+          ],
+        };
+        const response = await fetch(url, {
+          headers: {
+            "Content-Type": "application/json",
+          },
+          method: "POST",
+          body: JSON.stringify(requestBody),
+        });
+        const result = await response.json();
+        message = {
+          ok: true,
+          shortUrl: result?.shortUrl,
+        };
+      } catch (error) {
+        console.error(error);
+        message = {
+          ok: false,
+          error: `${error}`,
+        };
+      } finally {
+        io.to(data.socketConnectionId).emit("generateQREventMessage", {
+          ...message,
+        });
+      }
+    });
     socket.on("presentationEvent", (data) => {
       console.log("socket presentationEvent:", data);
       io.to(data.ref).emit("presentationEventMessage", data);
@@ -32,14 +82,6 @@ app.prepare().then(() => {
       console.log("A client disconnected");
     });
   });
-
-  const PORT = process.env.NEXT_PUBLIC_PORT
-    ? Number(process.env.NEXT_PUBLIC_PORT)
-    : 3000;
-
-  const PUBLIC_BASE_URL = process.env.NEXT_PUBLIC_BASE_URL
-    ? `${process.env.NEXT_PUBLIC_BASE_URL}:${PORT}`
-    : `http://localhost:${PORT}`;
 
   server.listen(PORT, (err) => {
     if (err) throw err;
